# shared/openapi/leyo-swimming-v1.yaml
openapi: 3.0.3
info:
  title: leyoSwimming API
  description: 游泳私教预约平台后端 API 契约（登录注册 + 套餐购买订单退款批次）
  version: 1.1.0
  contact:
    name: leyoSwimming Team
...
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /api/order/pay:
    post:
      tags:
        - Payment
      summary: 支付套餐订单
      description: 发起支付。错误码：ORDER_NOT_FOUND, ORDER_NOT_PAYABLE, ORDER_EXPIRED, INVALID_CHANNEL
      operationId: payOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderPayRequest'
      responses:
        '200':
          description: 支付结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderPayResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/payment/mock-callback:
    post:
      tags:
        - Payment
      summary: 支付 Mock 回调
      description: 模拟第三方支付回调，用于开发/测试环境。
      operationId: mockPaymentCallback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentMockCallbackRequest'
      responses:
        '200':
          description: 回调结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePaymentMockCallbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/package/list:
    post:
      tags:
        - Package
      summary: 正价套餐列表查询
      description: 游客可浏览平台标准套餐列表。
      operationId: listPackages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/coach/package/list:
    post:
      tags:
        - Coach
      summary: 教练关联套餐列表查询
      description: 查询某教练可售的标准套餐与自定义课包配置。错误码：COACH_NOT_FOUND
      operationId: listCoachPackages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/package/detail:
    post:
      tags:
        - Package
      summary: 套餐详情查询
      description: 查询套餐详情及适用教练。错误码：PACKAGE_NOT_FOUND
      operationId: getPackageDetail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/user/package/list:
    post:
      tags:
        - Package
      summary: 我的套餐列表
      description: 按状态分组查询当前学员的套餐。
      operationId: listUserPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/coach/package/detail:
    post:
      tags:
        - Coach
      summary: 套餐详情（教练/学员共用）
      description: 查询某个用户套餐的详情与使用记录。错误码：FORBIDDEN, PACKAGE_NOT_FOUND
      operationId: getUserPackageDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUserPackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/coach/student/package/list:
    post:
      tags:
        - Coach
      summary: 教练查看学员套餐列表
      description: 查询某学员在当前教练名下的套餐。错误码：NOT_ASSOCIATED_STUDENT
      operationId: listCoachStudentPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachStudentPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCoachStudentPackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/package/list:
    post:
      tags:
        - Admin Package
      summary: 管理员查看用户套餐列表
      description: 后台管理用户套餐分页查询。错误码：ADMIN_PERMISSION_DENIED
      operationId: listAdminPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/package/detail:
    post:
      tags:
        - Admin Package
      summary: 管理员查看用户套餐详情
      description: 查询单个用户套餐的完整信息。错误码：ADMIN_PERMISSION_DENIED, PACKAGE_NOT_FOUND
      operationId: getAdminPackageDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/package-template/list:
    post:
      tags:
        - Admin Package Template
      summary: 管理员查看套餐模板列表
      description: 查询标准套餐模板分页列表。错误码：FORBIDDEN
      operationId: listAdminPackageTemplates
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageTemplateListRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageTemplateListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/package-template/detail:
    post:
      tags:
        - Admin Package Template
      summary: 管理员查看套餐模板详情
      description: 查询套餐模板详情。错误码：FORBIDDEN, TEMPLATE_NOT_FOUND
      operationId: getAdminPackageTemplateDetail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageTemplateDetailRequest'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponsePackageTemplateDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/package-template/custom-config:
    post:
      tags:
        - Admin Package Template
      summary: 管理员配置自定义课包规则
      description: 设置自定义课时上下限等规则。错误码：FORBIDDEN, INVALID_CUSTOM_PACKAGE_CONFIG
      operationId: configAdminCustomPackage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPackageTemplateCustomConfigRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiResponseVoid'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
...
    PaginationRequest:
      type: object
      properties:
        page:
          type: integer
          default: 1
          minimum: 1
        pageSize:
          type: integer
          default: 10
          minimum: 1
          maximum: 50
...
    OrderPayRequest:
      type: object
      required:
        - orderId
        - channel
      properties:
        orderId:
          type: integer
        channel:
          type: integer
          description: 0=微信Mock，1=支付宝Mock
    OrderPayResponse:
      type: object
      properties:
        paymentId:
          type: integer
        channelTradeNo:
          type: string
        status:
          type: string
    ApiResponseOrderPayResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderPayResponse'
    PaymentMockCallbackRequest:
      type: object
      required:
        - channel
        - orderId
        - channelTradeNo
        - amount
        - success
      properties:
        channel:
          type: integer
        orderId:
          type: integer
        channelTradeNo:
          type: string
        amount:
          type: integer
        success:
          type: boolean
    PaymentMockCallbackResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
    ApiResponsePaymentMockCallbackResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PaymentMockCallbackResponse'
...
    PackageListRequest:
      $ref: '#/components/schemas/PaginationRequest'
    PackageListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        packageMode:
          type: string
        hours:
          type: integer
        price:
          type: integer
        originalPrice:
          type: integer
        teachingType:
          type: string
        durationMinutes:
          type: integer
        validityDays:
          type: integer
    PackageListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PackageListItem'
    ApiResponsePackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PackageListPageResponse'
    CoachPackageListRequest:
      type: object
      required:
        - coachId
      properties:
        coachId:
          type: integer
    CoachStandardPackage:
      type: object
      properties:
        id:
          type: integer
        hours:
          type: integer
        price:
          type: integer
        validityDays:
          type: integer
    CoachPackageListResponse:
      type: object
      properties:
        coachId:
          type: integer
        coachStatus:
          type: integer
        referencePrice:
          type: integer
        standardPackages:
          type: array
          items:
            $ref: '#/components/schemas/CoachStandardPackage'
        customPackageEnabled:
          type: boolean
        customHoursMin:
          type: integer
        customHoursMax:
          type: integer
    ApiResponseCoachPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachPackageListResponse'
    PackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
        coachId:
          type: integer
    CoachBrief:
      type: object
      properties:
        coachId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: string
    PackageDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        packageMode:
          type: string
        hours:
          type: integer
        price:
          type: integer
        originalPrice:
          type: integer
        teachingType:
          type: string
        durationMinutes:
          type: integer
        validityDays:
          type: integer
        refundPolicySummary:
          type: string
        applicableCoaches:
          type: array
          items:
            $ref: '#/components/schemas/CoachBrief'
    ApiResponsePackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PackageDetailResponse'
...
    UserPackageListRequest:
      type: object
      properties:
        status:
          type: string
    UserPackageItem:
      type: object
      properties:
        packageId:
          type: integer
        coachId:
          type: integer
        coachName:
          type: string
        coachAvatar:
          type: string
        packageName:
          type: string
        packageMode:
          type: string
        teachingType:
          type: string
        durationMinutes:
          type: integer
        validDays:
          type: integer
        totalHours:
          type: integer
        consumedHours:
          type: integer
        availableHours:
          type: integer
        reservedHours:
          type: integer
        paidAmount:
          type: integer
        status:
          type: string
        expireAt:
          type: string
          format: date-time
        refundEnabled:
          type: boolean
        refundRatio:
          type: number
        refundValidDays:
          type: integer
        frozenReason:
          type: string
    UserPackageSummary:
      type: object
      properties:
        totalHours:
          type: integer
        consumedHours:
          type: integer
        availableHours:
          type: integer
    UserPackageGroup:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserPackageItem'
        summary:
          $ref: '#/components/schemas/UserPackageSummary'
    UserPackageListResponse:
      type: object
      properties:
        active:
          $ref: '#/components/schemas/UserPackageGroup'
        exhausted:
          $ref: '#/components/schemas/UserPackageGroup'
        expired:
          $ref: '#/components/schemas/UserPackageGroup'
        refunded:
          $ref: '#/components/schemas/UserPackageGroup'
        frozen:
          $ref: '#/components/schemas/UserPackageGroup'
    ApiResponseUserPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserPackageListResponse'
    UserPackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
    PackageUsageRecord:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        consumedHours:
          type: integer
        consumedAt:
          type: string
          format: date-time
    UserPackageDetailResponse:
      type: object
      properties:
        packageId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        packageMode:
          type: string
        status:
          type: string
        totalHours:
          type: integer
        availableHours:
          type: integer
        reservedHours:
          type: integer
        consumedHours:
          type: integer
        expireAt:
          type: string
          format: date-time
        usageRecords:
          type: array
          items:
            $ref: '#/components/schemas/PackageUsageRecord'
    ApiResponseUserPackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserPackageDetailResponse'
...
    CoachStudentPackageListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          required:
            - studentId
          properties:
            studentId:
              type: integer
    CoachStudentPackageItem:
      type: object
      properties:
        packageId:
          type: integer
        packageName:
          type: string
        packageMode:
          type: string
        status:
          type: string
        statusLabel:
          type: string
        validStart:
          type: string
          format: date-time
        validEnd:
          type: string
          format: date-time
        remainingHours:
          type: integer
    CoachStudentPackageListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        packages:
          type: array
          items:
            $ref: '#/components/schemas/CoachStudentPackageItem'
    ApiResponseCoachStudentPackageListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CoachStudentPackageListPageResponse'
    AdminApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
    AdminApiResponseVoid:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              type: object
              nullable: true
...
    AdminPackageListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            status:
              type: string
            packageMode:
              type: string
            keyword:
              type: string
            expireAtStart:
              type: string
              format: date-time
            expireAtEnd:
              type: string
              format: date-time
    AdminPackageListItem:
      type: object
      properties:
        packageId:
          type: integer
        packageNo:
          type: string
        userId:
          type: integer
        userName:
          type: string
        coachId:
          type: integer
        coachName:
          type: string
        packageMode:
          type: string
        teachingType:
          type: string
        status:
          type: string
        availableCount:
          type: integer
        reservedCount:
          type: integer
        totalHours:
          type: integer
        expireAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    AdminPackageListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminPackageListItem'
    AdminApiResponsePackageListResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageListPageResponse'
    AdminPackageDetailRequest:
      type: object
      required:
        - packageId
      properties:
        packageId:
          type: integer
    AdminPackageDetailResponse:
      type: object
      properties:
        packageId:
          type: integer
        packageNo:
          type: string
        userId:
          type: integer
        userName:
          type: string
        coachId:
          type: integer
        coachName:
          type: string
        packageMode:
          type: string
        teachingType:
          type: string
        strokeIds:
          type: array
          items:
            type: integer
        totalHours:
          type: integer
        consumedCount:
          type: integer
        reservedCount:
          type: integer
        availableCount:
          type: integer
        paidAmount:
          type: integer
        originalPrice:
          type: integer
        refundEnabled:
          type: boolean
        refundRatio:
          type: number
        refundValidDays:
          type: integer
        status:
          type: string
        frozenReason:
          type: string
        extendReason:
          type: string
        expireAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        order:
          type: object
        bookingRecords:
          type: array
          items:
            type: object
        auditLogs:
          type: array
          items:
            type: object
    AdminApiResponsePackageDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageDetailResponse'
...
    AdminPackageTemplateListRequest:
      allOf:
        - $ref: '#/components/schemas/PaginationRequest'
        - type: object
          properties:
            coachId:
              type: integer
            status:
              type: string
    AdminPackageTemplateListItem:
      type: object
      properties:
        packageTemplateId:
          type: integer
        name:
          type: string
        packageMode:
          type: string
        teachingType:
          type: string
        coachIds:
          type: array
          items:
            type: integer
        coachNames:
          type: array
          items:
            type: string
        totalHours:
          type: integer
        durationMinutes:
          type: integer
        validDays:
          type: integer
        originalPrice:
          type: integer
        price:
          type: integer
        refundEnabled:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminPackageTemplateListPageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminPackageTemplateListItem'
    AdminApiResponsePackageTemplateListResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageTemplateListPageResponse'
...
    AdminPackageTemplateDetailRequest:
      type: object
      required:
        - packageTemplateId
      properties:
        packageTemplateId:
          type: integer
    AdminPackageTemplateDetailResponse:
      type: object
      properties:
        packageTemplateId:
          type: integer
        name:
          type: string
        packageMode:
          type: string
        teachingType:
          type: string
        coachIds:
          type: array
          items:
            type: integer
        coachList:
          type: array
          items:
            $ref: '#/components/schemas/CoachBrief'
        strokeIds:
          type: array
          items:
            type: integer
        totalHours:
          type: integer
        durationMinutes:
          type: integer
        validDays:
          type: integer
        originalPrice:
          type: integer
        price:
          type: integer
        refundEnabled:
          type: boolean
        refundRatio:
          type: number
        refundValidDays:
          type: integer
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminApiResponsePackageTemplateDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminPackageTemplateDetailResponse'
    AdminPackageTemplateCustomConfigRequest:
      type: object
      required:
        - minHours
        - maxHours
        - defaultValidDays
      properties:
        minHours:
          type: integer
        maxHours:
          type: integer
        defaultValidDays:
          type: integer
        unitPriceFloor:
          type: integer
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100001
            message: 请求参数错误
            data: null
    Unauthorized:
      description: 未授权或 token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 200001
            message: 登录已过期，请重新登录
            data: null
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 100002
            message: 资源不存在
            data: null
    Forbidden:
      description: 权限禁止或账号被禁用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 300001
            message: 账号已被禁用，请联系超级管理员
            data: null
...